More vram ticking

+++ src/vram.h
@@ -22,7 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 static const u32 vramsizes[] = { 64, 128, 256, 384, 512, 1024, 1536, 2048, 4096 };
 static const u32 vramelements = sizeof(vramsizes) / sizeof(u32);
 
-void initvram(const u32 size, const u32 edge);
+void initvram(const u64 size, const u32 edge);
 void allocbuf(const u32 id, const u32 size);
 void touchbuf(const u32 id);
 void freevram();
